Implement a decent GUI for selecting paste hints.
Even though the syntax for supplying a url with a mimetype hint (.ext) or a highlighting hint (/language) is pretty simple, it's still pretty daunting to browse through all of the /language ones.

There's still a quirk where the "no highlighting hints" message shows when the toggle switch is initially turned off.